backend/requirements.txt
# FastAPI and server
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0

# Auth (Story 1.3)
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.0
httpx>=0.27.0

# Secrets (Story 1.5)
google-cloud-secret-manager>=2.20.0

# Existing pipeline dependencies (shared)
requests>=2.31.0
gspread>=6.0.0
python-dotenv>=1.0.0
python-dateutil>=2.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
